[hlsearch-] move highlight code to features/
